Building. BUT END OF TIME for this branch. The effort to get it back working is  not worth it. Some of the  main problems: we have no longer the temporary tables lying around. Without it, we need to parse the LQP elements. That is part of another plugin and I do not see the effort being worth it

+++ src/cost_model_calibration_lib/cost_model_calibration_table_generator.cpp
@@ -63,7 +63,7 @@ void CostModelCalibrationTableGenerator::generate_calibration_tables() const {
   for (const auto& column_spec : _configuration.columns) {
     //column_specifications.emplace_back(ColumnSpecification(ColumnDataDistribution::make_uniform_config(0.0, column_spec.distinct_value_count),
 //		                        column_spec.data_type, column_spec.encoding, column_spec.column_name));
-    column_specifications.emplace_back(ColumnDataDistribution::make_uniform_config(0.0, column_spec.distinct_value_count),
+    column_specifications.emplace_back(ColumnDataDistribution::make_uniform_config(0.0, static_cast<double>(column_spec.distinct_value_count)),
 		                       column_spec.data_type, column_spec.encoding, column_spec.column_name);
   }
